Add files via upload
Constants are now float instead of double
Changed the way opponents are found
Finished wizard class
Created a way to deflect the correct damage

+++ Hero.java
@@ -50,7 +50,7 @@ public abstract class Hero implements Visitor, Visitable {
         }
     }
 
-    abstract double getLandModifier(LandType land);
+    abstract float getLandModifier(LandType land);
 
     public abstract void takeDmg(Visitor hero, LandType land);
